refactor tests/token -> tests/accounts
